MFB: Fixed bug #44333 (SEGFAULT when using mysql_pconnect() with client_flags)

+++ ext/mysql/php_mysql.c
@@ -18,7 +18,7 @@
    +----------------------------------------------------------------------+
 */
  
-/* $Id: php_mysql.c,v 1.213.2.6.2.16.2.11 2008/02/23 17:06:21 helly Exp $ */
+/* $Id: php_mysql.c,v 1.213.2.6.2.16.2.12 2008/03/04 22:29:29 felipe Exp $ */
 
 /* TODO:
  *
@@ -604,7 +604,7 @@ static void php_mysql_do_connect(INTERNAL_FUNCTION_PARAMETERS, int persistent)
 	int  user_len, passwd_len, host_len;
 	char *hashed_details=NULL;
 	int hashed_details_length, port = MYSQL_PORT;
-	int client_flags = 0;
+	long client_flags = 0;
 	php_mysql_conn *mysql=NULL;
 #if MYSQL_VERSION_ID <= 32230
 	void (*handler) (int);
